Fixing more bugs in the DUCT\MCTS player.

+++ src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/strategies/backpropagation/StandardBackpropagation.java
@@ -1,7 +1,7 @@
 package org.ggp.base.player.gamer.statemachine.MCTS.manager.strategies.backpropagation;
 
-import org.ggp.base.player.gamer.statemachine.MCTS.manager.treestructure.DUCTMove;
 import org.ggp.base.player.gamer.statemachine.MCTS.manager.treestructure.DUCTJointMove;
+import org.ggp.base.player.gamer.statemachine.MCTS.manager.treestructure.DUCTMove;
 import org.ggp.base.player.gamer.statemachine.MCTS.manager.treestructure.InternalPropnetDUCTMCTreeNode;
 
 public class StandardBackpropagation implements BackpropagationStrategy {
@@ -20,7 +20,7 @@ public class StandardBackpropagation implements BackpropagationStrategy {
 		int[] moveIndices = ductJointMove.getMovesIndices();
 
 		for(int i = 0; i < moves.length; i++){
-			moves[i][moveIndices[i]].incrementScoreSum(goals[i]);
+			moves[i][moveIndices[i]].incrementScoreSum((double) goals[i]);
 			moves[i][moveIndices[i]].incrementVisits();
 		}
